Ac_Admin_Manager_* improvements
+   Ac_Admin_Processing::$stayOn support (allows processing to address themselves)
+   Ac_Admin_Manager: ability to address manager JS controller from HTML elements
    with AvanControllers.findManager(element)   
+   Ac_Admin_Manager: ability to pass arbitrary parameters to 
    ManagerController.executeProcessing
+   Ajs_Util::makeQuery() can return associative array 
    with parameter names and string values
*   Removed unnecessary comment headers from some PHP files
*   Fixed many .htaccess files to make them both Apache2.2 and Apache2.4 compatible
+   Ac_Sql_Part can receive value from "value" prototype key
*   Ac_Model_Data::EVENT_ON_CHECK is now invoked after doOnCheck()
*   Ac_Form_Control presentation-related fixes and improvements

+++ Avancore/classes/Ac/Util/Lock.php
@@ -310,7 +310,7 @@ class Ac_Util_Lock extends Ac_Prototyped {
     
     function release($force = false) {
         $this->check(true);
-        if (($force || $this->lockStatus !== self::STATUS_OTHER) && $this->lockStatus !== self::STATUS_NONE) {
+        if (($force || $this->lockStatus !== self::STATUS_OTHER) && $this->lockStatus !== self::STATUS_NONE && $this->lockStatus !== self::STATUS_DELETED) {
             unlink($this->getPath());
         }
         return true;
